Convert py files to .ipynb for zero-shot
