+++ data/predictions/mistral7b/zero_shot/scores/ar_AE.jsonl
@@ -1,0 +1,6 @@
+{
+    "correct_instances": 13,
+    "total_instances": 722,
+    "comet_score": 0.6360196538496545,
+    "meta_score": 1.8005540166204987
+}
