+++ data/predictions/mistral7b/zero_shot/scores/de_DE.jsonl
@@ -1,0 +1,6 @@
+{
+    "correct_instances": 160,
+    "total_instances": 731,
+    "comet_score": 0.841678890826438,
+    "meta_score": 21.88782489740082
+}
