+++ data/predictions/mistral7b/zero_shot/scores/es_ES.jsonl
@@ -1,0 +1,6 @@
+{
+    "correct_instances": 231,
+    "total_instances": 739,
+    "comet_score": 0.8894736120762134,
+    "meta_score": 31.25845737483085
+}
