+++ data/predictions/mistral7b/zero_shot/scores/fr_FR.jsonl
@@ -1,0 +1,6 @@
+{
+    "correct_instances": 191,
+    "total_instances": 724,
+    "comet_score": 0.8442315421381049,
+    "meta_score": 26.381215469613263
+}
