+++ data/predictions/mistral7b/zero_shot/scores/it_IT.jsonl
@@ -1,0 +1,6 @@
+{
+    "correct_instances": 183,
+    "total_instances": 730,
+    "comet_score": 0.8693913776580602,
+    "meta_score": 25.068493150684933
+}
