+++ data/predictions/mistral7b/zero_shot/scores/ja_JP.jsonl
@@ -1,0 +1,6 @@
+{
+    "correct_instances": 23,
+    "total_instances": 723,
+    "comet_score": 0.7962425504894837,
+    "meta_score": 3.18118948824343
+}
