+++ data/predictions/mistral7b/zero_shot/scores/ko_KR.jsonl
@@ -1,0 +1,6 @@
+{
+    "correct_instances": 17,
+    "total_instances": 745,
+    "comet_score": 0.7929205839665944,
+    "meta_score": 2.2818791946308723
+}
