+++ data/predictions/mistral7b/zero_shot/scores/th_TH.jsonl
@@ -1,0 +1,6 @@
+{
+    "correct_instances": 2,
+    "total_instances": 710,
+    "comet_score": 0.4654504248373945,
+    "meta_score": 0.28169014084507044
+}
