+++ data/predictions/mistral7b/zero_shot/scores/tr_TR.jsonl
@@ -1,0 +1,6 @@
+{
+    "correct_instances": 74,
+    "total_instances": 732,
+    "comet_score": 0.7147966869623283,
+    "meta_score": 10.10928961748634
+}
