+++ data/predictions/mistral7b/zero_shot/scores/zh_TW.jsonl
@@ -1,0 +1,6 @@
+{
+    "correct_instances": 29,
+    "total_instances": 722,
+    "comet_score": 0.7770461272912673,
+    "meta_score": 4.016620498614958
+}
